refactor: enhance dataset processing logic
